tested basic operations and running

+++ src/controllers/admin.controller.js
@@ -14,46 +14,84 @@ async function deleteUser(req,res){
         res.status(400).json(errormessage.formatZodError(result.error)).end();
     }else{
         try{
-            const reqorder = await adminReqModel.findOne({_id:req.body.id})
-            console.log("Located delete-account-request to be deleted");
-            
-            await orderModel.deleteOne({_id:reqorder.objectId});
+            const reqorder = await adminReqModel.findOne({_id:req.body.id});
 
-            await adminReqModel.deleteOne({_id:req.body.id})
-            console.log("User deleted. Email verification follows");
-            res.status(200).json("User has been deleted succesfully...").end()
-        } catch(e){
-            res.status(500).json(e).end()
-        }   
+            if(reqorder){
+                console.log("Delete-account request found!");
+
+
+                await userModel.deleteOne({_id:reqorder.objectId});
+
+                await adminReqModel.deleteOne({_id:req.body.id})
+                console.log("User deleted. Email verification follows");
+    
+                res.status(200).json("User Account has been deleted succesfully...").end();
+            }else{
+                res.json("The requested delete-account order does not exist!").end();
+            };
+            }catch(err){
+                res.status(500).json(err).end();
+            }   
     }
 }
 
 async function deleteOrder(req,res){
-    const result = validator.deleteOrderValidator.safeParse(req.body)
+
+    const result = validator.deleteUserValidator.safeParse(req.body)
+
+    console.log(`validated:`.concat(result.success))
 
     if(!result.success){
         res.status(400).json(errormessage.formatZodError(result.error)).end();
-    } else{
+    }else{
         try{
-            const reqorder = await adminReqModel.findOne({_id: req.body.id})
+            const reqorder = await adminReqModel.findOne({_id:req.body.id});
+
             if(reqorder){
-                console.log("Located delete-order-request to be deleted");
-            } else{
-                console.log("cannot find the delete request");
-                res.status(400).json("Cannot order with such id").end()
-            }
+                console.log("Delete-order request found!");
+
+
+                await orderModel.deleteOne({_id:reqorder.objectId});
+
+                await adminReqModel.deleteOne({_id:req.body.id})
+                console.log("Order deleted. Email verification follows");
+    
+                res.status(200).json("Order has been deleted succesfully...").end();
+            }else{
+                res.json("The requested delete-order request does not exist!").end();
+            };
+            }catch(err){
+                res.status(500).json(err).end();
+            }   
+    }
+}
+
+// async function deleteOrder(req,res){
+//     const result = validator.deleteOrderValidator.safeParse(req.body)
+
+//     if(!result.success){
+//         res.status(400).json(errormessage.formatZodError(result.error)).end();
+//     } else{
+//         try{
+//             const reqorder = await adminReqModel.findOne({_id: req.body.id})
+//             if(reqorder){
+//                 console.log("Located delete-order-request to be deleted");
+//             } else{
+//                 console.log("cannot find the delete request");
+//                 res.status(400).json("Cannot order with such id").end()
+//             }
             
 
-            await userModel.deleteOne({_id:reqorder.objectId});
+//             await userModel.deleteOne({_id:reqorder.objectId});
 
-            await adminReqModel.deleteOne({_id:req.body.id})
-            console.log("Order deleted. Email verification follows");
-            res.status(200).json("Order has been deleted succesfully...").end()
-        } catch(err){
-            res.status(500).json(err).end()
-        }
-    }
-};
+//             await adminReqModel.deleteOne({_id:req.body.id})
+//             console.log("Order deleted. Email verification follows");
+//             res.status(200).json("Order has been deleted succesfully...").end()
+//         } catch(err){
+//             res.status(500).json(err).end()
+//         }
+//     }
+// };
 
 async function getAllUsers(){
     try{
